old/instagram_analyzer_with_api/requirements.txt
instaloader
Pillow
requests
python-dotenv 